[PATCH] Fixes for busybox init/shell

+++ debian/initscripts/etc/init.d/mountall.sh
@@ -33,7 +33,7 @@ then
 	rm -f $INITCTL
 	mknod -m 600 $INITCTL p
 fi
-kill -USR1 1
+[ "`readlink /sbin/init`" != "/bin/busybox" ] && kill -USR1 1
 
 #
 # Execute swapon command again, in case we want to swap to
